this was only needed before I added redux
